% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutenv.r
\name{perm.env}
\alias{perm.env}
\title{Permutation Test for Hierarchical Partitioning for Canonical Correspondence Analysis and Redundancy Analysis}
\usage{
perm.env(object, Y, X, type = "RDA", nper = 999)
}
\arguments{
\item{object}{one object from rdaenvpart}

\item{Y}{Community data matrix.}

\item{X}{Constraining matrix less than 12 columns, typically of environmental variables.}

\item{type}{the Constrained ordination: RDA or CCA, default "RDA"}
}
\value{
a list containing

\item{R2}{a dataframe for unadjusted R-squared for individual environmental variables and p-value.}

\item{adjR2}{a dataframe for adjusted R-squared for individual environmental variables and p-value.}
}
\description{
This function performs permutation test for the individual explain percentage of each environmental variable for Canonical Correspondence Analysis and Redundancy Analysis,
applying the hierarchy algorithm of Chevan and Sutherland (1991) .
}
\examples{
require(vegan)
data(varespec)
data(varechem)
mod=rdaenvpart(varespec,varechem[,c("Al","P","K")],pieplot = "tv",type="RDA")
perm.env(mod,varespec,varechem[,c("Al","P","K")],type="RDA",nper=999)
mod2=rdaenvpart(varespec,varechem[,c("Al","P","K")],pieplot = "tv",type="CCA")
perm.env(mod2,varespec,varechem[,c("Al","P","K")],type="CCA",nper=999)
}
\references{
Chevan, A. and Sutherland, M. 1991. Hierarchical Partitioning. The American Statistician 45:90~96
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
