% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdaenvpart.R
\name{rdaenvpart}
\alias{rdaenvpart}
\title{Hierarchical Partitioning for Canonical Correspondence Analysis and Redundancy Analysis}
\usage{
rdaenvpart(Y, X, type = "RDA", pieplot = "tv")
}
\arguments{
\item{Y}{Community data matrix.}

\item{X}{Constraining matrix less than 12 columns, typically of environmental variables.}

\item{type}{the Constrained ordination: RDA or CCA, default "RDA"}

\item{pieplot}{a pieplot each variable is plotted expressed as percentage of total variation (pieplot="tv") or total explained variation (pieplot="tev").}
}
\value{
a list containing

\item{R2}{unadjusted R-squared for RDA or CCA with all environmental variables.}

\item{gfs}{a data frame or vector listing all combinations of individual variables in the first column in ascending order, and the corresponding goodness of fit measure for the model using those variables.}

\item{IJ.R2}{a data frame of I, the individual and J the joint contribution for each individual variable (based on unadjusted R-squared).}

\item{I.perc}{a data frame of I as a percentage of total explained variation.}

\item{adj.R2}{adjusted R-squared for RDA or CCA with all environmental variables.}

\item{IJ.adjR2}{the individual contribution for each individual variable (based on adjusted R-squared).}
}
\description{
This function calculates the individual explain percentage of each environmental variable for Canonical Correspondence Analysis and Redundancy Analysis,
applying the hierarchy algorithm of Chevan and Sutherland (1991) .
}
\examples{
require(vegan)
data(varespec)
data(varechem)
rdaenvpart(varespec,varechem[,c("Al","P","K")],pieplot = "tv",type="RDA")
rdaenvpart(varespec,varechem[,c("Al","P","K")],pieplot = "tev",type="RDA")
rdaenvpart(varespec,varechem[,c("Al","P","K")],pieplot = "tv",type="CCA")
rdaenvpart(varespec,varechem[,c("Al","P","K")],pieplot = "tev",type="CCA")
}
\references{
Chevan, A. and Sutherland, M. 1991. Hierarchical Partitioning. The American Statistician 45:90~96
}
\author{
{Jiangshan Lai} \email{lai@ibcas.ac.cn}
}
